List of dataset objects

The class data.series is the blueprint of S4 objects that store a list of dataset objects.

Usage

# S4 method for data.series
show(object)

Arguments

object

an object of class data.series

Slots

series

(numeric) returns the number of dataset objects stored.

days

(numeric) returns the length of the simulated data in days common to all dataset objects stored. The default value is 60.

model

(character) returns a character string, either 'MPIN' or 'adjPIN'.

layers

(numeric) returns the number of information layers in all dataset objects stored. It takes the value 1 for the adjusted PIN model, i.e. when model takes the value 'adjPIN'.

datasets

(list) returns the list of the dataset objects stored.

restrictions

(list) returns a binary list that contains the set of parameter restrictions on the original AdjPIN model in the estimated AdjPIN model. The restrictions are imposed equality constraints on model parameters. If the value of the parameter restricted is the empty list (list()), then the model has no restrictions, and the estimated model is the unrestricted, i.e., the original AdjPIN model. If not empty, the list contains one or multiple of the following four elements {theta, mu, eps, d}. For instance, If theta is set to TRUE, then the estimated model has assumed the equality of the probability of liquidity shocks in no-information, and information days, i.e., \(\theta\)=\(\theta'\). If any of the remaining rate elements {mu, eps, d} is equal to TRUE, (say mu=TRUE), then the estimated model imposed equality of the concerned parameter on the buy side, and on the sell side (\(\mu\)b=\(\mu\)s). If more than one element is equal to TRUE, then the restrictions are combined. For instance, if the slot restrictions contains list(theta=TRUE, eps=TRUE, d=TRUE), then the estimated AdjPIN model has three restrictions \(\theta\)=\(\theta'\), \(\epsilon\)b=\(\epsilon\)s, and \(\Delta\)b=\(\Delta\)s, i.e., it has been estimated with just 7 parameters, in comparison to 10 in the original unrestricted model. [i] This slot only concerns datasets generated by the function generatedata_adjpin().

warnings

(numeric) returns numbers referring to the warning errors caused by a conflict between the different arguments used to call the function generatedata_mpin().

runningtime

(numeric) returns the running time of the data simulation in seconds.