AdjPIN estimation results

The class estimate.adjpin is a blueprint of the S4 objects that store the results of the estimation of the AdjPIN model using adjpin().

Usage

# S4 method for estimate.adjpin
show(object)

Arguments

object

(estimate.adjpin-class)

Slots

success

(logical) takes the value TRUE when the estimation has succeeded, FALSE otherwise.

errorMessage

(character) contains an error message if the estimation of the AdjPIN model has failed, and is empty otherwise.

convergent.sets

(numeric) returns the number of initial parameter sets, for which the likelihood maximization converged.

method

(character) contains a reference to the estimation method: "ECM" for expectation-conditional maximization algorithm and '"ML"' for standard maximum likelihood estimation.

factorization

(character) contains a reference to the factorization of the likelihood function used: "GE"for the factorization in Ersan and Ghachem (2022b) , and "NONE" for the original likelihood function in Duarte and Young (2009) .

restrictions

(list) returns a binary list that contains the set of parameter restrictions on the original AdjPIN model in the estimated AdjPIN model. The restrictions are imposed equality constraints on model parameters. If the value of the parameter restricted is the empty list (list()), then the model has no restrictions, and the estimated model is the unrestricted, i.e., the original AdjPIN model. If not empty, the list contains one or multiple of the following four elements {theta, mu, eps, d}. For instance, If theta is set to TRUE, then the estimated model has assumed the equality of the probability of liquidity shocks in no-information, and information days, i.e., \(\theta\)=\(\theta'\). If any of the remaining rate elements {mu, eps, d} is equal to TRUE, (say mu=TRUE), then the estimated model imposed equality of the concerned parameter on the buy side, and on the sell side (\(\mu\)b=\(\mu\)s). If more than one element is equal to TRUE, then the restrictions are combined. For instance, if the slot restrictions contains list(theta=TRUE, eps=TRUE, d=TRUE), then the estimated AdjPIN model has three restrictions \(\theta\)=\(\theta'\), \(\epsilon\)b=\(\epsilon\)s, and \(\Delta\)b=\(\Delta\)s, i.e., it has been estimated with just 7 parameters, in comparison to 10 in the original unrestricted model.

algorithm

(character) returns the implemented initial parameter set determination algorithm. "GE" is for Ersan and Ghachem (2022b) , "CL" is for Cheng and Lai (2021) , "RANDOM" for random initial parameter sets, and "CUSTOM" for custom initial parameter sets.

parameters

(numeric) returns the vector of the optimal maximum-likelihood estimates ( \(\alpha\), \(\delta\), \(\theta\), \(\theta'\), \(\epsilon\)b, \(\epsilon\)s, \(\mu\)b, \(\mu\)s, \(\Delta\)b, \(\Delta\)s).

likelihood

(numeric) returns the value (of the factorization) of the likelihood function, as in Ersan and Ghachem (2022b) , evaluated at the set of optimal parameters.

adjpin

(numeric) returns the value of the adjusted probability of informed trading (Duarte and Young 2009) .

psos

(numeric) returns the probability of symmetric order flow shock (Duarte and Young 2009) .

dataset

(dataframe) returns the dataset of buys and sells used in the estimation of the AdjPIN model.

initialsets

(dataframe) returns the initial parameter sets used in the estimation of AdjPIN model.

details

(dataframe) returns a dataframe containing the estimated parameters for each initial parameter set.

hyperparams

(list) returns the hyperparameters of the ECM algorithm, which are maxeval, and tolerance.

runningtime

(numeric) returns the running time of the AdjPIN estimation in seconds.