MPIN estimation results (ECM)

The class estimate.mpin.ecm is the blueprint of S4 objects that store the results of the estimation of the MPIN model using the Expectation-Conditional Maximization method, as implemented in the function mpin_ecm().

Usage

# S4 method for estimate.mpin.ecm
show(object)

selectModel(object, criterion)

# S4 method for estimate.mpin.ecm
selectModel(object, criterion)

getSummary(object)

# S4 method for estimate.mpin.ecm
getSummary(object)

Arguments

object

an object of class estimate.mpin.ecm.

criterion

a character string specifying the model selection criterion. criterion should take one of these values {"BIC", "AIC", "AWE"}. They stand for Bayesian Information Criterion, Akaike Information Criterion, and Approximate Weight of Evidence, respectively.

Functions

selectModel,estimate.mpin.ecm-method: returns the optimal model among the estimated models, i.e., the model having the lowest information criterion, provided by the user.

getSummary,estimate.mpin.ecm-method: returns a summary of the estimation of the MPIN model using the ECM algorithm for different values of the argument layers. For each estimation, the number of layers, the MPIN value, the log-likelihood value, as well as the values of the different information criteria, namely AIC, BIC and AWE are displayed.

Slots

success

(logical) returns the value TRUE when the estimation has succeeded, FALSE otherwise.

errorMessage

(character) returns an error message if the MPIN estimation has failed, and is empty otherwise.

convergent.sets

(numeric) returns the number of initial parameter sets at which the likelihood maximization converged.

method

(character) returns the method of estimation, and is equal to 'Expectation-Conditional Maximization Algorithm'.

layers

(numeric) returns the number of layers estimated by the Expectation-Conditional Maximization algorithm, or provided by the user.

optimal

(logical) returns whether the number of layers used for the estimation is provided by the user (optimal=FALSE), or determined by the ECM algorithm (optimal=TRUE).

parameters

(list) returns the list of the maximum likelihood estimates (\(\alpha\), \(\delta\), \(\mu\), \(\epsilon\)b, \(\epsilon\)s), where \(\alpha\), \(\delta\), and \(\mu\) are numeric vectors of length layers.

aggregates

(numeric) returns an aggregation of information layers' parameters alongside with \(\epsilon\)b and \(\epsilon\)s. The aggregated parameters are calculated as follows: \(\alpha_{agg} = \sum \alpha_j\)\(\alpha*= \sum \alpha\)j \(\delta_{agg} = \sum \alpha_j \times \delta_j\) \(\delta*= \sum \alpha\)j\(\delta\)j, and \(\mu_{agg} = \sum \alpha_j \times \mu_j\)\(\mu*= \sum \alpha\)j\(\mu\)j.

likelihood

(numeric) returns the value of the (log-)likelihood function evaluated at the optimal set of parameters.

mpinJ

(numeric) returns the values of the multilayer probability of informed trading per layer, calculated using the layer-specific estimated parameters.

mpin

(numeric) returns the global value of the multilayer probability of informed trading. It is the sum of the multilayer probabilities of informed trading per layer stored in the slot mpinJ.

mpin.goodbad

(list) returns a list containing a decomposition of MPIN into good-news, and bad-news MPIN components. The decomposition has been suggested for PIN measure in Brennan et al. (2016) . The list has four elements: mpinG, and mpinB are the global good-news, and bad-news components of MPIN, while mpinGj, and mpinBj are two vectors containing the good-news (bad-news) components of MPIN computed per layer.

dataset

(dataframe) returns the dataset of buys and sells used in the ECM estimation of the MPIN model.

initialsets

(dataframe) returns the initial parameter sets used in the ECM estimation of the MPIN model.

details

(dataframe) returns a dataframe containing the estimated parameters of the ECM method for each initial parameter set.

models

(list) returns the list of estimate.mpin.ecm objects storing the results of estimation using the function mpin_ecm() for different values of the argument layers. It returns NULL when the argument layers of the function mpin_ecm() take a specific value.

AIC

(numeric) returns the value of the Akaike Information Criterion (AIC).

BIC

(numeric) returns the value of the Bayesian Information Criterion (BIC).

AWE

(numeric) returns the value of the Approximate Weight of Evidence.

criterion

(character) returns the model selection criterion used to find the optimal estimate for the MPIN model. It takes one of these values 'BIC', 'AIC', 'AWE'; which stand for Bayesian Information Criterion, Akaike Information Criterion, and Approximate Weight of Evidence, respectively.

hyperparams

(list) returns the hyperparameters of the ECM algorithm, which are minalpha, maxeval, tolerance, and maxlayers. Check the details section of mpin_ecm() to know more about these parameters.

runningtime

(numeric) returns the running time of the estimation in seconds.